reorganize classes folder
Changes made (in order of having made them):

Moved rigidbody, hitbox, and engine files into physics folder within components folder

Moved animation, image, and spritesheet files into sprites folder within components folder

Moved button, slider, and text files nto ui folder within components folder

Moved game_object file into components folder

Renamed components folder to gameobject

Moved qtree into physics folder in gameobject folder

Moved camera into utils

Renamed classes folder to struct
