--- rubato/classes/components/__init__.py
@@ -1,14 +1,0 @@
-"""
-Holds all the various components for game objects.
-"""
-from .component import Component
-from .hitbox import Hitbox, Polygon, Circle, Rectangle
-from .rigidbody import RigidBody
-from .text import Text
-from .button import Button
-from .raster import Raster
-# from .slider import Slider # not implemented yet
-from .image import Image
-from .animation import Animation
-from .spritesheet import Spritesheet
-from .engine import Manifold, Engine
