+++ rubato/struct/__init__.py
@@ -1,10 +1,7 @@
 """
 The classes module holds all of the high-dependency modules relating to game objects.
 """
-from .camera import Camera
-from .components import *
-from .qtree import *
-from .game_object import GameObject
+from .gameobject import *
 from .group import Group
 from .scene import Scene
 from .scene_manager import SceneManager
