+++ rubato/struct/gameobject/__init__.py
@@ -1,0 +1,9 @@
+"""
+Holds all the various components for game objects.
+"""
+from .component import Component
+from .raster import Raster
+from .physics import *
+from .ui import *
+from .sprites import *
+from .game_object import GameObject
