+++ rubato/struct/gameobject/physics/__init__.py
@@ -1,0 +1,5 @@
+"""Holds all the physics related components"""
+from .hitbox import Hitbox, Polygon, Circle, Rectangle
+from .rigidbody import RigidBody
+from .engine import Manifold, Engine
+from .qtree import QTree
