+++ rubato/struct/gameobject/sprites/__init__.py
@@ -1,0 +1,4 @@
+"""Holds all image-related components."""
+from .image import Image
+from .animation import Animation
+from .spritesheet import Spritesheet
