+++ rubato/struct/gameobject/sprites/spritesheet.py
@@ -7,7 +7,7 @@ from typing import List
 import os
 
 from . import Animation
-from ... import Vector, get_path, Sprite
+from .... import Vector, get_path, Sprite
 
 
 class Spritesheet:
