+++ rubato/struct/gameobject/ui/__init__.py
@@ -1,0 +1,4 @@
+"""Holds the UI-related components."""
+from .text import Text
+from .button import Button
+# from .slider import Slider # not implemented yet
