+++ rubato/tests/group_test.py
@@ -1,10 +1,10 @@
 """Test the Group class."""
 from unittest.mock import Mock
 import pytest
-from rubato.classes.camera import Camera
-from rubato.classes.components.hitbox import Hitbox
-from rubato.classes.game_object import GameObject
-from rubato.classes.group import Group
+from rubato.utils.camera import Camera
+from rubato.struct.gameobject.physics.hitbox import Hitbox
+from rubato.struct.gameobject.game_object import GameObject
+from rubato.struct.group import Group
 from rubato.utils.error import Error
 # pylint: disable=redefined-outer-name
