+++ rubato/utils/__init__.py
@@ -17,3 +17,4 @@ from .sound import Sound
 from .quickselect import quickselect
 from .debug import Debug
 from .radio import Radio, Events
+from .camera import Camera
