+++ rubato/utils/camera.py
@@ -5,7 +5,7 @@ Items only render if their z-index is not more than that of the camera's.
 
 The current scene's camera can be accessed through :code:`Game.camera`.
 """
-from .. import Vector, Display, Math, Radio, Events
+from . import Vector, Display, Math, Radio, Events
 
 
 class Camera:
